FIX reverse import by creating an Component Object Model
